LayoutTests/fullscreen/video-controls-drag-expected.txt

This tests that the video element's controls are draggable in full screen. Press any key to continue.

EVENT(webkitfullscreenchange)
EXPECTED (endBox.left - startBox.left == '15') OK
EXPECTED (endBox.top - startBox.top == '15') OK
END OF TEST
